extract MarklightTextProcessor for style config
