+++ Marklight/MarklightTextProcessor.swift
@@ -1,0 +1,229 @@
+//
+//  MarklightTextProcessor
+//  Marklight
+//
+//  Created by Christian Tietze on 2017-07-20.
+//  Copyright © 2017 MacTeo. LICENSE for details.
+//
+
+#if os(iOS)
+    import UIKit
+#elseif os(macOS)
+    import AppKit
+#endif
+
+extension NSTextStorage {
+    func resetMarklightTextAttributes(textSize: CGFloat, range: NSRange) {
+        self.removeAttribute(NSForegroundColorAttributeName, range: range)
+        self.addAttribute(NSFontAttributeName, value: MarklightFont.systemFont(ofSize: textSize), range: range)
+        self.addAttribute(NSParagraphStyleAttributeName, value: NSParagraphStyle(), range: range)
+    }
+
+    #if os(iOS)
+    func invalidateTextSizeForWholeRange() {
+        let wholeRange = NSMakeRange(0, (self.string as NSString).length)
+        self.invalidateAttributes(in: wholeRange)
+        for layoutManager in self.layoutManagers {
+            layoutManager.invalidateDisplay(forCharacterRange: wholeRange)
+        }
+    }
+    #endif
+}
+
+public struct MarklightProcessingResult {
+    public let editedRange: NSRange
+    public let affectedRange: NSRange
+
+    public init(editedRange: NSRange, affectedRange: NSRange) {
+
+        self.editedRange = editedRange
+        self.affectedRange = affectedRange
+    }
+}
+
+open class MarklightTextProcessor {
+
+    let textStorage: NSTextStorage
+
+    open var string : String { return textStorage.string }
+
+    public init(textStorage: NSTextStorage) {
+        self.textStorage = textStorage
+
+        #if os(iOS)
+            observeTextSize()
+        #endif
+    }
+
+    // MARK: Syntax highlight customisation
+
+    /**
+     Color used to highlight markdown syntax. Default value is light grey.
+     */
+    open var syntaxColor = MarklightColor.lightGray
+
+    /**
+     Font used for blocks and inline code. Default value is *Menlo*.
+     */
+    open var codeFontName = "Menlo"
+
+    /**
+     `MarklightColor` used for blocks and inline code. Default value is dark grey.
+     */
+    open var codeColor = MarklightColor.darkGray
+
+    /**
+     Font used for quote blocks. Default value is *Menlo*.
+     */
+    open var quoteFontName = "Menlo"
+
+    /**
+     `MarklightColor` used for quote blocks. Default value is dark grey.
+     */
+    open var quoteColor = MarklightColor.darkGray
+
+    /**
+     Quote indentation in points. Default 20.
+     */
+    open var quoteIndendation : CGFloat = 20
+
+    /**
+     If the markdown syntax should be hidden or visible
+     */
+    open var hideSyntax = false
+
+
+    // MARK: Syntax highlighting
+
+    open func processEditing(
+        editedRange: NSRange
+        ) -> MarklightProcessingResult {
+
+        let editedAndAdjacentParagraphRange = self.editedAndAdjacentParagraphRange(editedRange: editedRange)
+        let editedParagraphRange = self.editedParagraphRange(editedRange: editedRange)
+
+        resetMarklightAttributes(range: editedAndAdjacentParagraphRange)
+
+        Marklight.syntaxColor = syntaxColor
+        Marklight.codeFontName = codeFontName
+        Marklight.codeColor = codeColor
+        Marklight.quoteFontName = quoteFontName
+        Marklight.quoteColor = quoteColor
+        Marklight.quoteIndendation = quoteIndendation
+        Marklight.textSize = textSize
+        Marklight.hideSyntax = hideSyntax
+
+        Marklight.processEditing(self.textStorage, affectedRange: editedAndAdjacentParagraphRange)
+
+        return MarklightProcessingResult(
+            editedRange: editedRange,
+            affectedRange: editedAndAdjacentParagraphRange)
+    }
+
+    fileprivate func editedParagraphRange(editedRange: NSRange) -> NSRange {
+        return (self.string as NSString).paragraphRange(for: editedRange)
+    }
+
+    fileprivate func editedAndAdjacentParagraphRange(editedRange: NSRange) -> NSRange {
+        let textStorageNSString = (self.string as NSString)
+        let editedParagraphRange = textStorageNSString.paragraphRange(for: editedRange)
+
+        let previousParagraphRange: NSRange
+        if editedParagraphRange.location > 0 {
+            previousParagraphRange = textStorageNSString.paragraphRange(at: editedParagraphRange.location - 1)
+        } else {
+            previousParagraphRange = NSRange(location: editedParagraphRange.location, length: 0)
+        }
+
+        let nextParagraphRange: NSRange
+        let locationAfterEditedParagraph = editedParagraphRange.location + editedParagraphRange.length
+        if locationAfterEditedParagraph < textStorageNSString.length {
+            nextParagraphRange = textStorageNSString.paragraphRange(at: locationAfterEditedParagraph + 1)
+        } else {
+            nextParagraphRange = NSRange.init(location: 0, length: 0)
+        }
+
+        return NSRange(
+            location: previousParagraphRange.location,
+            length: [previousParagraphRange, editedParagraphRange, nextParagraphRange].map { $0.length }.reduce(0, +))
+    }
+
+    fileprivate func resetMarklightAttributes(range: NSRange) {
+
+        textStorage.resetMarklightTextAttributes(
+            textSize: self.textSize,
+            range: range)
+    }
+
+    #if os(iOS)
+
+    /**
+     Internal method to register to notifications determined by dynamic type size
+     changes and redraw the attributed text with the appropriate text size.
+     Currently it works only after the user adds or removes some chars inside the
+     `UITextView`.
+     */
+    func observeTextSize() {
+        NotificationCenter.default.addObserver(forName: NSNotification.Name.UIContentSizeCategoryDidChange, object: nil, queue: OperationQueue.main) { [weak self] (notification) -> Void in
+            self?.textStorage.invalidateTextSizeForWholeRange()
+        }
+    }
+
+    deinit {
+        NotificationCenter.default.removeObserver(self)
+    }
+
+    // MARK: Font Style Settings
+
+    /**
+     Dynamic type font text style, default `UIFontTextStyleBody`.
+
+     - see: [Text Styles](xcdoc://?url=developer.apple.com/library/ios/documentation/UIKit/Reference/UIFontDescriptor_Class/index.html#//apple_ref/doc/constant_group/Text_Styles)
+     */
+    open var fontTextStyle : String = UIFontTextStyle.body.rawValue
+
+    /// Text size measured in points.
+    fileprivate var textSize: CGFloat {
+        return MarklightFontDescriptor
+            .preferredFontDescriptor(withTextStyle: UIFontTextStyle(rawValue: self.fontTextStyleValidated))
+            .pointSize
+    }
+
+    // We are validating the user provided fontTextStyle `String` to match the
+    // system supported ones.
+    fileprivate var fontTextStyleValidated : String {
+
+        let supportedTextStyles: [String] = {
+
+            let baseStyles = [
+                UIFontTextStyle.headline.rawValue,
+                UIFontTextStyle.subheadline.rawValue,
+                UIFontTextStyle.body.rawValue,
+                UIFontTextStyle.footnote.rawValue,
+                UIFontTextStyle.caption1.rawValue,
+                UIFontTextStyle.caption2.rawValue
+            ]
+
+            guard #available(iOS 9.0, *) else { return baseStyles }
+
+            return baseStyles.appending(contentsOf: [
+                UIFontTextStyle.title1.rawValue,
+                UIFontTextStyle.title2.rawValue,
+                UIFontTextStyle.title3.rawValue,
+                UIFontTextStyle.callout.rawValue
+                ])
+        }()
+        
+        guard supportedTextStyles.contains(self.fontTextStyle) else {
+            return UIFontTextStyle.body.rawValue
+        }
+        
+        return self.fontTextStyle
+    }
+
+    #elseif os(macOS)
+
+    open var textSize: CGFloat = NSFont.systemFontSize()
+
+    #endif
+}
