+++ Marklight/MarklightTextStorage.swift
@@ -11,9 +11,15 @@
     import AppKit
 #endif
 
-extension NSString {
-    func paragraphRange(at location: Int) -> NSRange {
-        return paragraphRange(for: NSRange(location: location, length: 0))
+extension NSLayoutManager {
+    func processEditing(for textStorage: NSTextStorage, processingResult: MarklightProcessingResult) {
+
+        self.processEditing(
+            for: textStorage,
+            edited: .editedAttributes,
+            range: processingResult.editedRange,
+            changeInLength: 0,
+            invalidatedRange: processingResult.affectedRange)
     }
 }
 
@@ -78,47 +84,13 @@ extension NSString {
 
 open class MarklightTextStorage: NSTextStorage {
 
+    open weak var marklightTextProcessor: MarklightTextProcessor?
+
     /// Delegate from this class cluster to a regular `NSTextStorage` instance
     /// because it does some additional performance optimizations 
     /// over `NSMutableAttributedString`.
     fileprivate let imp = NSTextStorage()
-    
-    // MARK: Syntax highlight customisation
-    
-    /**
-    Color used to highlight markdown syntax. Default value is light grey.
-    */
-    open var syntaxColor = MarklightColor.lightGray
-    
-    /**
-     Font used for blocks and inline code. Default value is *Menlo*.
-     */
-    open var codeFontName = "Menlo"
-    
-    /**
-     `MarklightColor` used for blocks and inline code. Default value is dark grey.
-     */
-    open var codeColor = MarklightColor.darkGray
-    
-    /**
-     Font used for quote blocks. Default value is *Menlo*.
-     */
-    open var quoteFontName = "Menlo"
-    
-    /**
-     `MarklightColor` used for quote blocks. Default value is dark grey.
-     */
-    open var quoteColor = MarklightColor.darkGray
-    
-    /**
-     Quote indentation in points. Default 20.
-     */
-    open var quoteIndendation : CGFloat = 20
-   
-    /**
-     If the markdown syntax should be hidden or visible
-     */
-    open var hideSyntax = false
+
         
     // MARK: Syntax highlighting
 
@@ -128,8 +100,7 @@ open class MarklightTextStorage: NSTextStorage {
 
     /**
     To customise the appearance of the markdown syntax highlights you should
-     subclass this class (or create your own direct `NSTextStorage` subclass)
-     and set the customisations in this method implementation. Sends out
+     subclass `MarklightTextProcessor`. Sends out
      `-textStorage:willProcessEditing`, fixes the attributes, sends out
      `-textStorage:didProcessEditing`, and notifies the layout managers of
      change with the
@@ -144,81 +115,27 @@ open class MarklightTextStorage: NSTextStorage {
         self.isBusyProcessing = true
         defer { self.isBusyProcessing = false }
 
-        let editedAndAdjacentParagraphRange = self.editedAndAdjacentParagraphRange
-        let editedParagraphRange = self.editedParagraphRange
-        defer {
-            // Include surrounding paragraphs in layout manager's styling pass
-            // after finishing the real edit. Mostly needed for Setex headings.
-            layoutManagers.first?.processEditing(
-                for: self,
-                edited: .editedAttributes,
-                range: editedRange,
-                changeInLength: 0,
-                invalidatedRange: editedAndAdjacentParagraphRange)
+        if let marklightTextProcessor = marklightTextProcessor {
+            let processingResult = marklightTextProcessor.processEditing(editedRange: self.editedRange)
+
+            defer {
+                // Include surrounding paragraphs in layout manager's styling pass
+                // after finishing the real edit. Mostly needed for Setex headings.
+                self.layoutManagers.forEach {
+                    $0.processEditing(
+                        for: self,
+                        processingResult: processingResult)
+                }
+            }
         }
-
-        resetMarklightAttributes(range: editedAndAdjacentParagraphRange)
-//        removeParagraphAttributes()
-//        removeWholeAttributes()
-
-        Marklight.syntaxColor = syntaxColor
-        Marklight.codeFontName = codeFontName
-        Marklight.codeColor = codeColor
-        Marklight.quoteFontName = quoteFontName
-        Marklight.quoteColor = quoteColor
-        Marklight.quoteIndendation = quoteIndendation
-        Marklight.textSize = textSize
-        Marklight.hideSyntax = hideSyntax
-        
-        Marklight.processEditing(self, affectedRange: editedAndAdjacentParagraphRange)
 
         super.processEditing()
     }
 
-    fileprivate var editedParagraphRange: NSRange {
-        return (self.string as NSString).paragraphRange(for: self.editedRange)
-    }
-
-    fileprivate var editedAndAdjacentParagraphRange: NSRange {
-        let textStorageNSString = (self.string as NSString)
-        let editedParagraphRange = textStorageNSString.paragraphRange(for: self.editedRange)
-
-        let previousParagraphRange: NSRange
-        if editedParagraphRange.location > 0 {
-            previousParagraphRange = textStorageNSString.paragraphRange(at: editedParagraphRange.location - 1)
-        } else {
-            previousParagraphRange = NSRange(location: editedParagraphRange.location, length: 0)
-        }
-
-        let nextParagraphRange: NSRange
-        let locationAfterEditedParagraph = editedParagraphRange.location + editedParagraphRange.length
-        if locationAfterEditedParagraph < textStorageNSString.length {
-            nextParagraphRange = textStorageNSString.paragraphRange(at: locationAfterEditedParagraph + 1)
-        } else {
-            nextParagraphRange = NSRange.init(location: 0, length: 0)
-        }
-
-        return NSRange(
-            location: previousParagraphRange.location,
-            length: [previousParagraphRange, editedParagraphRange, nextParagraphRange].map { $0.length }.reduce(0, +))
-    }
-
-    fileprivate func resetMarklightAttributes(range: NSRange) {
+    override func resetMarklightTextAttributes(textSize: CGFloat, range: NSRange) {
         imp.removeAttribute(NSForegroundColorAttributeName, range: range)
         imp.addAttribute(NSFontAttributeName, value: MarklightFont.systemFont(ofSize: textSize), range: range)
         imp.addAttribute(NSParagraphStyleAttributeName, value: NSParagraphStyle(), range: range)
-    }
-
-    // Remove every attribute the the paragraph containing the last edit.
-    fileprivate func removeParagraphAttributes() {
-        let paragraphRange = (string as NSString).paragraphRange(for: self.editedRange)
-        resetMarklightAttributes(range: paragraphRange)
-    }
-
-    // Remove every attribute to the whole text
-    fileprivate func removeWholeAttributes() {
-        let wholeRange = NSMakeRange(0, (self.string as NSString).length)
-        resetMarklightAttributes(range: wholeRange)
     }
 
 
@@ -313,102 +230,4 @@ open class MarklightTextStorage: NSTextStorage {
         edited([.editedAttributes], range: range, changeInLength: 0)
         endEditing()
     }
-
-    // MARK: - iOS-Only Font Text Style Support
-
-    #if os(iOS)
-
-    // MARK: Initialisers
-
-    /**
-     The designated initialiser. If you subclass `MarklightTextStorage`, you
-     must call the super implementation of this method.
-     */
-    override public init() {
-        super.init()
-        observeTextSize()
-    }
-
-    /**
-     The designated initialiser. If you subclass `MarklightTextStorage`, you must
-     call the super implementation of this method.
-     */
-    required public init?(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)
-        observeTextSize()
-    }
-
-    /**
-     Internal method to register to notifications determined by dynamic type size
-     changes and redraw the attributed text with the appropriate text size.
-     Currently it works only after the user adds or removes some chars inside the
-     `UITextView`.
-     */
-    func observeTextSize() {
-        NotificationCenter.default.addObserver(forName: NSNotification.Name.UIContentSizeCategoryDidChange, object: nil, queue: OperationQueue.main) { (notification) -> Void in
-            let wholeRange = NSMakeRange(0, (self.string as NSString).length)
-            self.invalidateAttributes(in: wholeRange)
-            for layoutManager in self.layoutManagers {
-                layoutManager.invalidateDisplay(forCharacterRange: wholeRange)
-            }
-        }
-    }
-
-    deinit {
-        NotificationCenter.default.removeObserver(self)
-    }
-
-    // MARK: Font Style Settings
-
-    /**
-     Dynamic type font text style, default `UIFontTextStyleBody`.
-
-     - see: [Text Styles](xcdoc://?url=developer.apple.com/library/ios/documentation/UIKit/Reference/UIFontDescriptor_Class/index.html#//apple_ref/doc/constant_group/Text_Styles)
-     */
-    open var fontTextStyle : String = UIFontTextStyle.body.rawValue
-
-    /// Text size measured in points.
-    fileprivate var textSize: CGFloat {
-        return MarklightFontDescriptor
-            .preferredFontDescriptor(withTextStyle: UIFontTextStyle(rawValue: self.fontTextStyleValidated))
-            .pointSize
-    }
-
-    // We are validating the user provided fontTextStyle `String` to match the
-    // system supported ones.
-    fileprivate var fontTextStyleValidated : String {
-
-        let supportedTextStyles: [String] = {
-
-            let baseStyles = [
-                UIFontTextStyle.headline.rawValue,
-                UIFontTextStyle.subheadline.rawValue,
-                UIFontTextStyle.body.rawValue,
-                UIFontTextStyle.footnote.rawValue,
-                UIFontTextStyle.caption1.rawValue,
-                UIFontTextStyle.caption2.rawValue
-            ]
-
-            guard #available(iOS 9.0, *) else { return baseStyles }
-
-            return baseStyles.appending(contentsOf: [
-                UIFontTextStyle.title1.rawValue,
-                UIFontTextStyle.title2.rawValue,
-                UIFontTextStyle.title3.rawValue,
-                UIFontTextStyle.callout.rawValue
-                ])
-        }()
-
-        guard supportedTextStyles.contains(self.fontTextStyle) else {
-            return UIFontTextStyle.body.rawValue
-        }
-        
-        return self.fontTextStyle
-    }
-
-    #elseif os(macOS)
-
-    open var textSize: CGFloat = NSFont.systemFontSize()
-
-    #endif
 }
